Now norm of the function works properly, but not the kappasqr. The DIAG is calculated only if ldiag==1.

+++ gpu_prelim/main.cu
@@ -3,6 +3,7 @@
 #include "cuda.h"
 #include "evolve.h"
 #include "model.h"
+#include <time.h>
 /*==============================================*/
 int main( void ){
   CRASH BUG, *dev_bug ; 
@@ -22,21 +23,24 @@ int main( void ){
   }
   printf( "#-I shall launch %d blocks, each with %d threads\n", Nblock, Nthread );
   /*------------------------------------------------*/
-
   set_crash(  &BUG, &dev_bug ); 
   alloc_chain( &PSI, &dev_psi );
   set_param( &PARAM, &dev_param ) ;
+  bool Exit= check_param( PARAM );
+  if (Exit){return 0;}
   int size_diag = pre_diag( &DIAG , &dev_diag, PARAM );
-  int size_psi = NN*sizeof(double);
+  // int size_psi = NN*sizeof(double);
   // pre_evolve(  ndim, "rnkt4" , &TT, &dev_tt, Nblock, Nthread  ) ;
   // pre_evolve(  ndim, "rnkt4" , &TT, &dev_tt, Nblock, Nthread  ) ;
   pre_evolve(  ndim, TimeScheme , &TT, &dev_tt, Nblock, Nthread  ) ;
   // setup initial configuration 
   initial_configuration( PSI, PARAM ) ;
   system("exec rm -rf data");
-  system("exec mkdir data");
+  system("exec mkdir data"); 
   // wPSI( PSI, TT.time ) ;   
   H2D( dev_psi, PSI, ndim );
+  clock_t timer=clock();
+  double ttimer = timer/CLOCKS_PER_SEC;
   printf( " #starting time evolution ...\n ");
   evolve( PSI, dev_psi, 
           &TT,  dev_tt,
@@ -45,6 +49,8 @@ int main( void ){
           BUG,  dev_bug, 
           Nblock, Nthread ) ;
   printf( "#... time evolution finished \n");
+  ttimer = clock()/CLOCKS_PER_SEC-ttimer;
+  printf("Total time taken %fs\n", ttimer);
   D2H( PSI, dev_psi, ndim );
   // cudaMemcpy( PSI, dev_psi, size_psi , cudaMemcpyDeviceToHost);
   wPSI( PSI, TT.time ) ;
